backend/requirements.txt
axios==0.4.0
Django==4.2.1
django-cors-headers==4.0.0
django-rest-framework==0.1.0
djoser==2.2.0